Create Remapping
In order to detect which new species are there for the individual, we will extract the sequences information from the .bam file header and compare to the species list in the original reference definition file. Then we need to count the new reads on the new species and filter the very low abundance ones.

We should update the individual references, rebuild and remap. We will instead, as a first approximation, add these new reads to the previous and redo the count matrix. The difference would be minimal, and expected only in the fraction of multiply mapped reads.